+++ packages/db/db-client-types.ts
@@ -13,6 +13,9 @@ export type UpsertUserInput = {
 }
 
 export type UpsertUserOutput = {
+  id: string
+  // whether the upsert inserted a new record (if falsy, it was updated)
+  inserted: boolean
   issuer: string
 }