+++ packages/website/tests/accountPayment.e2e.spec.ts
@@ -10,14 +10,21 @@ test.beforeEach(async ({ page }) => {
 });
 
 test.describe('/account/payment', () => {
-  test('redirects to /login/ when not authenticated', async ({ page }, testInfo) => {
-    await Promise.all([
-      page.waitForNavigation({
-        waitUntil: 'networkidle',
-      }),
-      page.goto('/account/payment'),
-    ]);
-    await expect(page).toHaveURL('/login/');
+  test('redirects through login and back when not initially authenticated', async ({ page }, testInfo) => {
+    const accountPaymentPathname = '/account/payment';
+    // try to go to page that requires authn
+    await page.goto(accountPaymentPathname, { waitUntil: 'networkidle' });
+    // wait for redirect to a Log in page
+    await page.locator('text=Log in').waitFor();
+    const pageUrl = new URL(page.url());
+    expect(pageUrl.pathname).toEqual('/login/');
+    expect(pageUrl.searchParams.get('redirect_uri')).toEqual(accountPaymentPathname);
+    // fill login
+    await LoginTester().login(page, {
+      email: MAGIC_SUCCESS_EMAIL,
+    });
+    // should be back at our initial target destination
+    expect(new URL(page.url()).pathname).toEqual(accountPaymentPathname + '/');
     await page.screenshot({
       fullPage: true,
       path: await E2EScreenshotPath(testInfo, `accountPayment-noauth`),
@@ -27,35 +34,40 @@ test.describe('/account/payment', () => {
     page.on('pageerror', err => {
       console.error('pageerror', err);
     });
-    const goToPayment = (page: Page) =>
-      Promise.all([
-        page.waitForNavigation({
-          waitUntil: 'networkidle',
-        }),
-        page.goto('/account/payment'),
-      ]);
-    await goToPayment(page);
-    await LoginTester().login(page, {
+    const tester = LoginTester();
+    await page.goto(tester.loginHref);
+    await tester.login(page, {
       email: MAGIC_SUCCESS_EMAIL,
     });
-    // @todo - this should redirect you back to where you wanted to go: /account/payment
-    await expect(page).toHaveURL('/account/');
-    await goToPayment(page);
+    await page.goto('/account/payment/');
     await expect(page).toHaveURL('/account/payment/');
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    const currentPlanTitle = page.locator('[data-testid="currentPlan.title"]');
+    // TODO - this won't work in CI because the env.subscriptions used is a mock that
+    // only persists in-memory, which doesn't work in our miniflare setup.
+    // This check does persist when tested locally using StripeSubscriptionsService.
+    // One way of making this work would be to have the mockSubscriptionsService read/write from disk
+    // expect(await currentPlanTitle.innerText()).toBe('Free');
     await page.screenshot({
       fullPage: true,
       path: await E2EScreenshotPath(testInfo, `accountPayment`),
     });
   });
   test('can enter credit card details', async ({ page }) => {
-    await LoginTester().login(page, { email: MAGIC_SUCCESS_EMAIL });
+    const tester = LoginTester();
+    await page.goto(tester.loginHref);
+    await tester.login(page, { email: MAGIC_SUCCESS_EMAIL });
     await page.goto(AccountPaymentTester().url);
     await AccountPaymentTester().fillCreditCardDetails(page);
     await AccountPaymentTester().clickAddCardButton(page);
   });
 });
 
-function LoginTester() {
+function LoginTester({
+  loginHref = '/login/',
+}: {
+  loginHref?: string;
+} = {}) {
   async function fillLoginForm(page: Page, email: string) {
     await page.locator('.login-email').fill(email);
   }
@@ -67,20 +79,15 @@ function LoginTester() {
   async function login(
     page: Page,
     {
-      url = '/login/',
       email,
     }: {
-      url?: string;
       email: string;
     }
   ) {
-    if (new URL(page.url()).pathname !== url) {
-      await page.goto(url);
-    }
     await fillLoginForm(page, email);
     await submitLoginForm(page);
   }
-  return { login, submitLoginForm };
+  return { login, loginHref, submitLoginForm };
 }
 
 function AccountPaymentTester() {
@@ -91,6 +98,7 @@ function AccountPaymentTester() {
       await stripeFrame.locator('[placeholder="Card number"]').fill('4242424242424242');
       await stripeFrame.locator('[placeholder="MM / YY"]').fill('04/30');
       await stripeFrame.locator('[placeholder="CVC"]').fill('242');
+      await stripeFrame.locator('[placeholder="ZIP"]').fill('42424');
     },
     async clickAddCardButton(page: Page) {
       await page.locator('text=Payment MethodsAdd Card >> button').click();